requirements.txt
numpy<2
cudf-cu12 
cuml-cu12
jax[cuda12]
cupy-cuda12x
pandas
scipy
matplotlib
seaborn
scikit-learn
pysindy
pydmd
ipykernel
jupyter
ipython
torch
setuptools
wheel